fix(Handler): no internal instance cache
For orders this is not needed. the internal instance cache causes more problems
due to the mix between uuid and id cache. the internal cache on uuid / id was therefore removed.
Orders have the data freshly fetched from the db anyway. therefore unnecessary.
In addition, a wrong order object was repeatedly retrieved when sending mail.