fix(Basket): improve basket id handling and user check in Basket.php
Added a conditional check in src/QUI/ERP/Order/Basket/Basket.php to convert basketId to actual id
from user's session if it is a boolean value. This fixes potential issues that may arise with
non-integer basketIds.

Also, introduced a safeguard check before saving to ensure there is a valid User. If not, the save
function is exited early to prevent potential errors or undesired actions.