fix(phpstan): update abstractorder and other files for consistent function behaviour
Updated AbstractOrder.php to return null instead of false in getShippingStatus() when the shipping
status is not set. This makes the method's behaviour more consistent.

Updated Search.php to add a condition check for class existence before retrieving shipping status.
This ensures that the code will not fail if the class is not defined.

Modified DataLayer.php to add a condition check for class existence before retrieving shipping.

Added @phpstan-ignore-next-line in Utils.php to silence phpstan for the next line of code.

Related: quiqqer/order#172