chore(phpstan): ignore phpstan errors due to missing classes
Ignored various PHPStan errors in 'phpstan-baseline.neon' file due to missing classes and methods.
Many of these errors were related to the ERP, Shipping, and Invoice modules. This fix suppresses
the error reporting to ensure smooth operation of the rest of the code base. However, it's
recommended to address these missing dependencies for a long-term solution.