fix: add version compatibility in Mailer creation
The changes in `src/QUI/ERP/Order/Mail.php` determines whether the 'getMailer' method accepts
parameter or not. This is achieved by using reflection. If the method accepts parameters (which
means it's a newer version), the original code execution proceeds. If it doesn't (the older
version), 'getMailer' gets invoked without any parameters. This way, compatibility is ensured
between the older and newer versions.

Related: pcsg/buero#525