refactor: update mailer instantiation in Order Mail
This commit refactors the way we get the Mailer in the sendOrderConfirmationMail method of Order
Mail. Instead of directly calling the getMailer method on QUI::getMailManager(), we first store the
MailManager in a variable and then call getMailer on it. This change is aimed at improving code
readability and maintainability.